split rank type questions into 2 classes
