+++ src/main/resources/feedbackQuestionRankEditFormOptionFragment.html
@@ -5,8 +5,8 @@
                 id="${Const.ParamsNames.FEEDBACK_QUESTION_RANKOPTION}-${i}-${questionNumber}"
                 class="constSumOptionTextBox" value="${rankOptionValue}">
         <span class="input-group-btn">
-            <button class="btn btn-default removeOptionLink" type="button" id="constSumRemoveOptionLink" 
-                    onclick="removeConstSumOption(${i},${questionNumber})"
+            <button class="btn btn-default removeOptionLink" type="button" id="rankRemoveOptionLink" 
+                    onclick="removeRankOption(${i},${questionNumber})"
                     style="display:none" tabindex="-1">
                 <span class="glyphicon glyphicon-remove">
                 </span>
