+++ src/main/resources/feedbackQuestionRankRecipientsEditFormTemplate.html
@@ -2,9 +2,6 @@
     <br>
     <div class="col-sm-6" id="rankOptionTable-${questionNumber}" style="display:none">          
         
-        <input type="hidden" name="${Const.ParamsNames.FEEDBACK_QUESTION_RANKTORECIPIENTS}"
-            id="${Const.ParamsNames.FEEDBACK_QUESTION_RANKTORECIPIENTS}-${questionNumber}" 
-            value="true">
     </div>
     <div class="col-sm-6">
         <input type="checkbox" name="${Const.ParamsNames.FEEDBACK_QUESTION_RANKISDUPLICATESALLOWED}"
